# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Oficina Mecanica - API (monolito)
  description: |
    API de gerenciamento da oficina: ordens de servico, catalogos, estoque e
    clientes.

    Autenticacao NAO faz parte deste contrato. `POST /auth/login` e
    `POST /auth/register` sao servidos por uma funcao Lambda cujo contrato vive
    em `oficina-mecanica-serverless/docs/openapi.yaml`. Este servico apenas
    valida o token emitido la; obtenha um em `/api/auth/login` antes de chamar
    qualquer rota protegida.
  version: 3.0.0
  license:
    name: Uso academico
servers:
  - url: /api
    description: Atras do CloudFront/API Gateway
  - url: /
    description: Servidor atual (execucao local direta)
...
security:
  - BearerAuth: []
...
paths:
...
  /services:
    post:
      operationId: postServices
      tags: [Services]
      summary: Criar servico da oficina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopServiceCreateRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Servico criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: getServices
      tags: [Services]
      summary: Listar servicos da oficina
      parameters:
        - name: title
          in: query
          description: Busca parcial por titulo (case insensitive)
          schema:
            type: string
        - name: active
          in: query
          description: Filtrar por status ativo
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        <<: *protectedResponses
        "200":
          description: Lista paginada de servicos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /services/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getServicesId
      tags: [Services]
      summary: Buscar servico por ID
      responses:
        <<: *protectedResponses
        "200":
          description: Servico encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putServicesId
      tags: [Services]
      summary: Atualizar servico (parcial)
      description: Apenas os campos fornecidos sao alterados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopServiceUpdateRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: Servico atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteServicesId
      tags: [Services]
      summary: Deletar ou desativar servico
      description: |
        Hard delete se nao ha vinculos com OS.
        Se ha vinculos, desativa (soft delete) e retorna 200.
      responses:
        <<: *protectedResponses
        "204":
          description: Servico deletado (hard delete)
        "200":
          description: Servico desativado (soft delete)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: service has existing work order links and was deactivated
                  service:
                    $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /services/avg-execution-time:
    get:
      operationId: getServicesAvgExecutionTime
      tags: [Reports]
      summary: Tempo medio de execucao por servico
      description: |
        Retorna o tempo medio real de execucao para cada tipo de servico,
        calculado a partir de servicos finalizados em ordens de servico.
      parameters:
        - name: from
          in: query
          description: Data inicio (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-01-01"
        - name: to
          in: query
          description: Data fim (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-12-31"
        - name: technician_id
          in: query
          description: Filtrar pelo tecnico atribuido (UUID). O alias legado `technicianId` tambem e aceito temporariamente.
          schema:
            type: string
            format: uuid
      responses:
        <<: *protectedResponses
        "200":
          description: Lista com tempos medios
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AvgExecutionTimeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Work Orders ───────────────────────────────────────
...
  /work-orders/{id}/services:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: postWorkOrdersIdServices
      tags: [Work Orders]
      summary: Adicionar servicos a uma OS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AddServiceRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Servicos adicionados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrderService"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: Status da OS nao permite adicionar servicos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /work-orders/{id}/services/{wosId}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
      - name: wosId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID do servico na OS (work_order_service)
    delete:
      operationId: deleteWorkOrdersIdServicesWosId
      tags: [Work Orders]
      summary: Remover servico de uma OS
      responses:
        <<: *protectedResponses
        "204":
          description: Servico removido
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Status da OS nao permite remocao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Numero da pagina (default 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Itens por pagina (default 10)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: Requisicao invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid id"
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "resource not found"
    Conflict:
      description: "Conflito (ex: titulo duplicado)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "service title already exists"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    WorkshopServiceCreateRequest:
      type: object
      required: [title, price_cents, estimated_time_minutes]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 150
          example: Troca de Oleo
        description:
          type: string
          maxLength: 500
          example: Troca completa de oleo do motor com filtro
        price_cents:
          type: integer
          minimum: 1
          description: Preco em centavos (ex. 5000 = R$50.00)
          example: 5000
        estimated_time_minutes:
          type: integer
          minimum: 1
          example: 30
    WorkshopServiceUpdateRequest:
      type: object
      description: Pelo menos um campo deve ser fornecido.
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 150
        description:
          type: string
          maxLength: 500
        price_cents:
          type: integer
          minimum: 1
        estimated_time_minutes:
          type: integer
          minimum: 1
        active:
          type: boolean
    WorkshopServiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price_cents:
          type: integer
        estimated_time_minutes:
          type: integer
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkshopServiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WorkshopServiceResponse"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    AvgExecutionTimeResponse:
      type: object
      properties:
        service_id:
          type: string
          format: uuid
        title:
          type: string
          example: Troca de Oleo
        estimated_time_minutes:
          type: integer
          example: 30
        avg_real_time_minutes:
          type: number
          format: double
          example: 25.5
        execution_count:
          type: integer
          example: 3
        difference_minutes:
          type: number
          format: double
          example: -4.5

    # ── Work Order ────────────────────────────────────
...
    WorkOrderService:
      type: object
      required: [id, work_order_id, service_id, service_title_snapshot, service_price_cents_snapshot, service_estimated_time_minutes_snapshot, approval_status, status, supplies, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        work_order_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        service_title_snapshot:
          type: string
        service_description_snapshot:
          type: string
          nullable: true
        service_price_cents_snapshot:
          type: integer
        service_estimated_time_minutes_snapshot:
          type: integer
          description: Tempo base estimado do servico em minutos (antes de eventuais acrescimos de prazo no orcamento).
        approval_status:
          type: string
          enum: [PENDENTE, APROVADO, REPROVADO]
        status:
          type: string
          enum: [PENDENTE, AGUARDANDO_APROVACAO, EM_EXECUCAO, FINALIZADO]
        supplies:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderServiceSupply"
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkOrderServiceSupply:
      type: object
      required: [id, work_order_service_id, supply_id, supply_title_snapshot, supply_price_cents_snapshot, supply_quantity, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        work_order_service_id:
          type: string
          format: uuid
        supply_id:
          type: string
          format: uuid
        supply_title_snapshot:
          type: string
        supply_price_cents_snapshot:
          type: integer
        supply_quantity:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AddServiceRequest:
      type: object
      required: [service_id]
      properties:
        service_id:
          type: string
          format: uuid
        estimated_time_minutes:
          type: integer
          description: Override do tempo estimado base em minutos (opcional).
...